+++ types/api/blocks.d.ts
@@ -1,5 +1,6 @@
 import {OutputData} from '../data-formats/output-data';
-import {BlockToolData, ToolConfig} from "../tools";
+import {BlockToolData, ToolConfig} from '../tools';
+import {BlockAPI} from './block';
 
 /**
  * Describes methods to manipulate with Editor`s blocks
@@ -49,7 +50,7 @@ export interface Blocks {
    * @param {number} index
    * @returns {HTMLElement}
    */
-  getBlockByIndex(index: number): HTMLElement;
+  getBlockByIndex(index: number): BlockAPI;
 
   /**
    * Returns current Block index
@@ -61,6 +62,8 @@ export interface Blocks {
    * Mark Block as stretched
    * @param {number} index - Block to mark
    * @param {boolean} status - stretch status
+   *
+   * @deprecated Use BlockAPI interface to stretch Blocks
    */
   stretchBlock(index: number, status?: boolean): void;